add(Repository): add a movies repository
